aiueo

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,24 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->simplePaginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -51,13 +62,28 @@ class TasksController extends Controller
             'status' => 'required|max:10',   // add
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;    // add
-        $task->save();
+
+        $request->user()->tasks()->create([
+        'content' => $request->content,
+        'status' => $request->status,    
+        ]);
+        
 
         return redirect('/');
     }
+    
+    /*public function store(Request $request)
+    {
+        $this->validate($request, [
+            'content' => 'required|max:191',
+        ]);
+
+        $request->user()->microposts()->create([
+            'content' => $request->content,
+        ]);
+
+        return redirect('/');
+    }*/
 
     /**
      * Display the specified resource.
@@ -70,7 +96,11 @@ class TasksController extends Controller
         $task = Task::find($id);
 
         return view('tasks.show', [
+            
             'task' => $task,
+/*            'id' => $task->id,
+            'content' => $task->content,
+            'status' => $task->status*/
         ]);
     }
 
@@ -86,6 +116,7 @@ class TasksController extends Controller
 
         return view('tasks.edit', [
             'task' => $task,
+            'status' => $task->status,
         ]);
     }
 
@@ -119,8 +150,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+
+        $task = \App\Task::find($id);
+
+/*var_dump($task);
+return;*/
+
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
